[PATCH] Fix ext3 transaction batching
ext3 transaction batching has been ineffective since the scheduler changes
forced us to replace the yield() with a schedule().

Using schedule_timeout(1) fixes it up again.  Benchmarking is positive with
wither a 1 or 10 millisecond delay in there, so there appears to be no need
to play around with HZ.

BKrev: 407afb27UdNGDeqHaggAGuKopt_XoA