Merge bk://bk.arm.linux.org.uk/linux-2.6-serial into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/04/11 19:08:55+01:00 rmk
[SERIAL] Add support for TI16C750 hardware flow control.

2004/04/10 12:00:44+01:00 rmk
[SERIAL] Add extra suspend/resume functionality to serial_cs.

This calls into the 8250 driver so that the serial port settings can
be saved and restored over a suspend/resume cycle.  Previous kernels
have assumed that the port will be re-opened after such an event,
which may not be the case.

2004/04/10 10:28:36+01:00 bjorn.helgaas
[SERIAL] HCDP IRQ fixup

Some pre-production firmware has incorrect GSI values in the
HCDP, which tells us where the serial console port is, so we
have to do the auto-IRQ thing after all.

2004/04/09 23:01:39+01:00 rmk
[SERIAL] Remove UPF_HUP_NOTIFY; this is no longer used.

2004/04/09 22:57:30+01:00 rmk
[SERIAL] Pass sa11x0 struct device through to tty_register_device.

2004/04/09 22:52:50+01:00 rmk
[SERIAL] Pass device pointer through to tty_register_device.

This allows drivers to pass their struct device through to
tty_register_device, which in turn allows sysfs to show which
device and driver owns the UART.

2004/04/09 22:33:45+01:00 rmk
[SERIAL] Don't try to free resources we didn't request.

2004/04/09 22:29:22+01:00 rmk
[SERIAL] Correct minor debugging format string error.

2004/04/09 22:23:30+01:00 rmk
[SERIAL] Remove some dead declarations.

2004/04/09 22:16:29+01:00 rmk
[SERIAL] Unuse old SERIAL_IO_xxx macros.

8250.c should be using the replacement UPIO_xxx macros instead.

2004/04/08 19:43:18+01:00 ben-linux
[ARM PATCH] 1807/1: S3C2410 - onboard serial

Patch from Ben Dooks

Serial driver for S3C2410 on board UARTs. Re-post of 1796/1

Includes BAST driver to register on-board 16550s.

BKrev: 407b24d3nDHnGLaq2qY3bw_o5usNWA