[PATCH] ppc64: Fix ibmveth.c compilation
From: Paul Mackerras <paulus@samba.org>

This patch changes PCI_DMA_TODEVICE to DMA_TO_DEVICE in a couple of
places in drivers/net/ibmveth.c, since it doesn't compile without this
change and it does compile with it.  It also reformats a couple of
over-long lines in the vicinity of the other changes.

BKrev: 407c14aemuU_0CwMO1wHjBqbJl9kRw