Merge redhat.com:/spare/repo/netdev-2.6/pcnet32 into redhat.com:/spare/repo/net-drivers-2.6
2004/04/13 14:06:34-04:00 brazilnut
[PATCH] pcnet32 fix hang/crash with loopback test

If the pcnet32 interface is not up, running the loopback test may hang or
crash the system.  This patch provided by Jim Lewis fixes that problem.
Tested on ia32 and ppc systems.

2004/04/06 07:57:37-04:00 brazilnut
[PATCH] pcnet32 add led blink capability

Please apply this patch to 2.6.5 to include the capability of blinking
the LED's for device identification.  Ethtool -p support.

2004/04/06 07:57:29-04:00 brazilnut
[PATCH] pcnet32 correct name display

This displays the device name (eth%d) instead of the device type
when registering the interrupt handler.

2004/04/06 07:57:21-04:00 brazilnut
[PATCH] pcnet32 all printk under netif_msg

Please apply the following patch to 2.6.5.

All printk's are now done under netif_msg debug level tests.

2004/04/06 07:57:12-04:00 brazilnut
[PATCH] pcnet32.c add support for 79C976

Please include this patch to add support for the 79C976 device to the
pcnet32 driver.

BKrev: 407c38770vthP8s3L5rzRaue4auRGQ