[PATCH] NMI watchdog Pentium M support
This adda nmi_watchdog=2 support to the Pentium M processor.  The P-M is
a P6 chip, but it shares some chipset logic with the Pentium 4, so it
requires this workaround to function.

Without this patch, NMI gets stuck after 1 count.  With it, the NMI
fires and breaks me out of UHCI-related hard lockups.

This patch is basically a modified version of the same patch for
oprofile. See the threaded discussion here:

	http://lkml.org/lkml/2004/2/12/181

BKrev: 407fe343uMM6rqg8Pvr7egZSkxP_vQ