[PATCH] ppc64: hugepage cleanup
From: David Gibson <david@gibson.dropbear.id.au>

This is a small cleanup to the PPC64 hugepage code.  It removes an
unhelpful function, removing some studlyCaps in the process.  It was
originally this way to match the normal page path, but that has all been
rewritten since.

BKrev: 4081cb87-XNdiXycggD_PjPBoaluZQ