Merge bk://bk.arm.linux.org.uk/linux-2.6-pcmcia into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/04/17 09:41:18+01:00 proski
[PCMCIA] Conversion to module_param

Patch from: Pavel Roskin

As it turns out, mixing MODULE_PARM and module_param in one module is
wrong.  The parameters specified in module_param are ignored.  I've just
posted a patch to LKML that will detect this condition and warn about it.

The new debugging code used the new-style module_param, which means that
all instances of MODULE_PARM should be converted.  The attached patch does
that.

An additional bonus is that module_param_array provides the number of
array elements.  This allowed me to change tcic.c and i82365.c to use
this number for IRQ list.  This change was tested with i82365.  If
"irq_list" is not specified, irq_list_count is 0.

I set all permissions to 0444 to be safe.  I think we have no secrets
from the users regarding those parameters.  If some parameters can be
changed safely at the runtime, the permissions could be changed to 0644.
I didn't examine how safe (and how useful) it would be, so it's 0444 for
now.

BKrev: 4081cf7f9dGNvTDfqY6m_PkJF5wc_Q