Merge bk://gkernel.bkbits.net/misc-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/04/19 12:54:45-04:00 bjorn.helgaas
[PATCH] Fix hw_random build on ia64

The following patch allows hw_random.c to build on ia64.  (The problem
was just that the VIA stuff has i386 assembly in it.  The current code
only probes for VIA on i386 anyway, so this patch just adds more ifdefs
so the VIA code is only built for i386.)

BKrev: 40842675XLAvkIcX6qa_5HzYhYM3Ow