[PATCH] ISDN Eicon driver: remove call to trap usermode helper
From: Armin Schindler <armin@melware.de>

This patch removes the function to call a usermode helper if the hardware
and its firmware has a problem.  This feature is already implemented in the
user utility and the driver does not need to care about it.

BKrev: 408df771-Rsi4vkIVuaiom7L969xIg