[PATCH] Fix overeager stack-expansion on ppc64
This fix is from Paul Mackerras and was applied in 2.4 sometime late
last year.

On ppc64, touching addresses between the highest other mapping and the
stack can cause the stack to be extended way, way down, rather than
causing a SEGV as you would expect.  This patch only allows the stack
mapping to be extended to cover addresses actually within the stack
(as determined by looking at the process's r1).  This fix is ported
from 2.4

This fixes failures on the LTP's shmdt01, munmap01 and munmap02 tests.

BKrev: 40910d2fOM11FFDJgi0kPihruAejww