[PATCH] cciss build fix
From: <mikem@beardog.cca.cpqcorp.net>

This patch fixes the linux/include/cciss_ioctl.h file.  When support for the
cciss big ioctl was added the stucture in the header was put in the wrong
place.  If an application includes the file it will fail to compile.

BKrev: 40910ebdFQqgrCSiHmbqD1GJflCilQ