[PATCH] ppc32: Updated boot fix
This fixes booting on some PPC32 machines, notably CHRP and powermac
machines.  This is a modified version of Tom Rini's patch that addresses
the concerns I had with it.

The problem was that the linker script was getting included in the list
of things that got put together to make some of the sorts of bootable
images that we produce.  This removes ld.script in cases where it wasn't
appropriate and changes the rules in others so that although we have the
dependency on ld.script, it doesn't get included in the list of things
to link.

BKrev: 4097af63Dijf_2UsCmUx3VV65osktg