Merge bk://bk.arm.linux.org.uk/linux-2.6-serial into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/05/05 00:17:47+01:00 rmk
[SERIAL] Remove unused variable.

2004/05/04 17:56:47+01:00 bjorn.helgaas
[SERIAL] default to serial console when possible

Patch from Bjorn Helgaas

This adds efi_uart_console_only() so we can default to using a serial
console if the EFI console path only contains UARTs.

2004/05/04 16:18:13+01:00 alex.williamson
[SERIAL] 8250_hcdp needs irq sharing

Patch from Alex Williamson

Here's a trivial patch that makes 8250_hcdp setup the correct flags
when IRQ sharing is enabled for serial ports.

The HCDP table tells us if the device is a PCI UART.  We can use this
to set the shared interrupt flag as well as program the interrupt with
the correct polarity/trigger (should get rid of "changing vector <x>
from IO-SAPIC-edge to IO-SAPIC-level" messages at bootup).  This also
allows non-PCI UARTs to be left un-shareable, which is likely much
more safe (edge triggered).

The bit that I'm keying on is still part of the older 1.0a HCDP spec,
so should be implemented (it was on all the boxes I tested).  If
there's firmware out there that doesn't set this bit or the interrupt
supported flag, the HCDP UART may run in polling mode, but should
still be functional.

2004/05/04 15:09:30+01:00 SKolodynski
[SERIAL] Add support for SBS Tech. Inc. PMC-OCTPRO and P-OCTAL cards.

Patch from Slawomir Kolodynski

BKrev: 4098404byPvvsS1ciVD9Y97-zn5E9A