Merge flint.arm.linux.org.uk:/usr/src/bk/linux-2.6-sharp into flint.arm.linux.org.uk:/usr/src/bk/linux-2.6-rmk
2004/05/09 15:35:01+01:00 elf
[ARM PATCH] 1818/1: lh7a40x #2 (3/7) doc

Patch from Marc Singer

Documentation for the Sharp-LH machines.


2004/05/09 15:31:40+01:00 elf
[ARM PATCH] 1817/1: lh7a40x #2 (2/7) core-include

Patch from Marc Singer

Include files for this updated lh7a40x patch set.  The changes in this
set from the previous are mostly cosmetic.  The memory macros were
reworked in order to be more similar to the other ARM versions.  The
previous versions produced the same results, but the forms are
slightly different.


2004/05/09 15:27:51+01:00 elf
[ARM PATCH] 1816/1: lh7a40x #2 (1/7) core

Patch from Marc Singer

Updated change set for the 2.6.5 kernel *and* for the April 8th arm
patch.  Also included are changes suggested by Russell that merge
several of the files in the mach- directory.  I have also endeavored
to remove all unnecessary whitespace additions.

Note that since I've found the cause of an annoying user-space crash,
I believe that this patch is OK.  The crash appears to have nothing to
do with the system setup.

BKrev: 409e46c9KWzI3ceXyTR_CUXQWIy9Cw